fix(gui): split panel headers into .cpp and fix gradient mode bugs
Move MixedFilamentColorMapPanel, MixedGradientSelector, and
MixedColorMatchHelpers implementations from header-only to proper
.cpp/.hpp pairs, and register them in CMakeLists.txt.

Fix duplicate wxEVT_TEXT binding on pattern control, clamp gradient
mode to exactly 2 filament rows, hide add/remove buttons in gradient
mode, replace fixed-pixel strip preview with Bresenham-scheduled
24-segment rendering for consistent cross-DPI behavior, and guard
against division-by-zero when the pattern vector is empty.
