+++ src/slic3r/CMakeLists.txt
@@ -314,10 +314,13 @@ set(SLIC3R_GUI_SOURCES
     GUI/AddColorMixDialog.cpp
     GUI/AddColorMixDialog.hpp
     GUI/MixedGradientSelector.hpp
+    GUI/MixedGradientSelector.cpp
     GUI/MixedColorMatchPanel.cpp
     GUI/MixedColorMatchPanel.hpp
     GUI/MixedColorMatchHelpers.hpp
+    GUI/MixedColorMatchHelpers.cpp
     GUI/MixedFilamentColorMapPanel.hpp
+    GUI/MixedFilamentColorMapPanel.cpp
     GUI/ObjColorDialog.cpp
     GUI/ObjColorDialog.hpp
     GUI/ObjectDataViewModel.cpp
